Test all modules with `hk.jit(mod)`.
We need to lift the restriction that params in apply are immutable, this is
because `hk.jit` needs to plumb parameter values from the result of calling
`jit` back into the frame.

I suspect there is room for improvement here, one obvious thing we could
optimize is that most parameters to `hk.jit` will be passed through (e.g. we
pass all parameters in but only a subset will be used). We could either add a
pass on jaxpr that would optimize the graph passed to XLA (e.g. remove any
`outvars` that are also `invars`). I haven't closely measured the performance
yet so will hold back from doing any optimization.

PiperOrigin-RevId: 293206502
Change-Id: I7db304b5fe206e7f547ad8efc41631ff3c9ec75f