Add `hk.to_dot(f)`, an experimental JAX+Haiku to dot tracer.
Introduces an (internal only) module hook mechanism, allowing callers to hook
all module method calls in for some scope and observe enter and exit (currently
does not pass input/output values). We use this to implement clustering on top
of our tracer.

The testing on the to_dot function is quite minimal for the moment, I want to
get this in early to get feedback and allow folks to use it. To make it
reasonable to test (I didn't want to try and string pattern match the dot
itself) I have extracted a graph IR that we can test on, the "rendering" code
for the IR->dot is trivial enough that I feel OK skipping detailed unit tests
here.

Finally I have tested that this works (as in produces a non-empty result and
does not crash) for all built-in Haiku modules with an integration test.

PiperOrigin-RevId: 300713874
Change-Id: Idc19bbd2f32615a616e5eb67bfc55acf876ec1b1