Add params filtering utilities to Haiku
Haiku transform returns an apply function taking as first argument a single params object containing all the variables used in the computation graph.
Often times we might want to apply transformations (such as gradients and jacobian computation) only wrt to a subset of these variables, for example corresponding to a specific sub-graph of interest, and this CL provides a minimal set functions that make these operations easier to perform in a few lines of code (demonstrated in the tests).
More specifically the CL includes
* a function to partition a params/state object according to a user specified predicate
* a function to fitler the params/state object using a user specified predicate
* a function to merge a number of input params/state objects.
Note that using a predicate is strictly more expressive than using a regex - the predicate can infact take the form regex.match.
In the test users can find reference implementation of a utility to apply a transformation to a function f, via an intermediate step where f is wrapped by an auxiliary function whose signature allows for controlling which params are passed as first arguments.

PiperOrigin-RevId: 301168486
Change-Id: If7f0205af1edf9c682454617488de5de0a3cd764