Updates ResetCore's should_reset signal shape handling.
With this change ResetCore will be able to handle should_reset inputs of shape (batch_size,) regardless of the rank of the state tensors. More generally, users will be able to pass both tensor and nested should_reset inputs, as long as they match the state nest structure, and their shape is a prefix of the shape of the corresponding state components.

PiperOrigin-RevId: 310142722
Change-Id: I0a72dfba1c53f53ff0d5b56b8cb2803794c188f1