Ensure that shim `hk` imports match public API.
We have started using a shim for `hk` in internal modules to make them easier to
read and fork. One mistake we have made a few times already is to not be
consistent with the actual public API, this test makes sure that at least top
level symbols are aligned in all internal modules.

PiperOrigin-RevId: 312968408
Change-Id: I0e059cf272489a373ba00815e900df0b00d7b50e