Add support for @dataclass modules.
Inspired by google/flax#208 I wondered if we could similarly offer support for
@dataclass to define modules. It seems we can :) One thing I wanted to maintain
was support for custom names. To do this in the examples below I add an optional
name property to the dataclasses (you can omit this if you prefer to take the
default name) and then we override it with the uniquified name as part of
`__post_init__`.

```python
@dataclass
class Linear(hk.Module):
  output_size: int
  name: Optional[str] = None

  def __call__(self, x):
    j, k = x.shape[-1], self.output_size
    w = hk.get_parameter("w", [j, k], init=jnp.ones)
    b = hk.get_parameter("b", [k], init=jnp.zeros)
    return x @ w + b

@dataclass
class MLP(hk.Module):
  output_sizes: Sequence[int]
  activation: Callable[[jnp.ndarray], jnp.ndarray] = jax.nn.relu
  name: Optional[str] = None

  def __call__(self, x):
    for i, output_size in enumerate(self.output_sizes):
      if i > 0:
        x = self.activation(x)
      x = Linear(output_size)(x)
    return x
```

PiperOrigin-RevId: 313392449
Change-Id: I4c30d34474bd78c567c4101958fc12c8c82f3efa