Create `hk.scan` and use it to implement `hk.dynamic_unroll`.
Omnistaging (google/jax#3370) is going to introduce a new, dynamic scoping
mechanism for JAX tracers. Currently we have been lucky with `scan` and
`dynamic_unroll`, since the function we have been scanning has typically only
been impure during `init` and the only side effect has been creating params with
an rng key captured from outside. While technically this is a violation of the
contract for scan (`f` was not a pure function), it has been working fine
(because `f` only has side effects that `scan` cannot observe). However..
omnistaging sees all and will trigger an error here (thankfully this has been
caught by our existing tests!!).

The fix here is to explicitly thread Haiku state in and out of the `scan` and
use this `hk.scan` to implement `dynamic_unroll`. I have tried to make this
more efficient by only threading `state` (e.g. batch norm averages) and `rng` in
and out of the `scan`. This is only safe to do since during `init` we unroll the
first step of the `scan`.

PiperOrigin-RevId: 318067893
Change-Id: I9844d46cfc282e08544ca0eae819940627552d28