Add original_dtype to ParamContext. This allows custom_getters to reason about the original call to `get_parameter` and coordinate better with their corresponding custom_creators.
Imagined use-case: we want all bfloat16 parameters to have an underlying float32 parameter. Therefore the custom_creator needs to know to override the bfloat16 dtype, but the custom_getter needs to know to re-cast the parameter to bfloat16 in the case that the parameter was originally intended to be bfloat16.

PiperOrigin-RevId: 323378467
Change-Id: I0e574b829d1183e7b2a8acab1606d4145f14a193