Add method interceptor API.
Method interceptors allow a convenient and robust way to monkey patch methods on
Haiku modules (builtin or custom). The design is similar to the custom getter
API for parameters in that interceptors stack and you are passed the next
interceptor to call (with the last "interceptor" being the original bound
method).

This API allows for other features such as mixed precision training to be built
quite trivially by allowing policy decisions (e.g. whether to cast a module's
parameters or use a custom creator/getter) to be taken on a per module method
basis.

PiperOrigin-RevId: 324208890
Change-Id: I2fbcb8fc8c63c02791d72e1bb16e896ffdcb4e28