Initial version of `hk.vmap(..)`.
Supports mapping modules leaving parameters/state unmapped. Morally this is like
`hk.BatchApply` in that it allows you to run a module with an additional input
dimension.

In the longer term we should offer control over whether parameters, state and
rng are mapped or not. This requires more careful thought and design, so users
in the mean time are advised to make use of `hk.experimental.lift` and
`jax.vmap`.

PiperOrigin-RevId: 335663135
Change-Id: Ie689c5f96df5487450cf14b2621e74cfdeb7bf7c