Add hk.running_init().
In general you should not need to gate behaviour of your module based on
whether you are running `init` or `apply`, but sometimes (e.g. when making use
of JAX control flow) this is required.

For example, if you want to use `hk.switch(..)` to pick between experts, when
we run your init function we need to ensure that params/state for all experts
are created (unconditionally) but during apply we want to conditionally apply
(and perhaps update the internal state) of only one of our experts.

In that case we need to have different behaviour in init vs. apply:

```python
experts = [hk.nets.ResNet50(10) for _ in range(5)]
x = jnp.ones([1, 224, 224, 3])
if hk.running_init():
  # During init unconditionally create params/state for all experts.
  for expert in experts:
    out = expert(x, is_training=True)
else:
  # During apply conditionally apply (and update) only one expert.
  index = jax.random.randint(hk.next_rng_key(), [], 0, len(experts) - 1)
  out = hk.switch(index, experts, x)
```
PiperOrigin-RevId: 336301736
Change-Id: I62a07d6e6b4fdd499bfb3b819c86a701b52878fb