Fix instance check for RNNCore under Python 3.8.
Under Python 3.8.6 our tests were failing with:

    AttributeError: type object 'RNNCore' has no attribute '_is_protocol'

This seems to be a difference between `typing_extensions` in <3.8 and `typing`.
The following is sufficient to reproduce the error without Haiku (using
`typing_extensions` in 3.6 this works fine):

    import abc
    from typing import Protocol

    class Foo(object, metaclass=type(Protocol)):
      @abc.abstractmethod
      def f(self):
        pass

    isinstance(object(), Foo)
    # Traceback (most recent call last):
    #   File "/tmp/test.py", line 9, in <module>
    #     isinstance(object(), Foo)
    #   File "/usr/lib/python3.8/typing.py", line 1016, in __instancecheck__
    #     if cls._is_protocol:
    # AttributeError: type object 'Foo' has no attribute '_is_protocol'

For now we will work around the issue, but I am not fully clear whether this is
a bug in typing, or whether what we are doing (extending ProtocolMeta but not
Protocol) is unsupported.

PiperOrigin-RevId: 351329027
Change-Id: I757d12f33d8bb4d98f5801ab0d9935ceef4ace8a