Add an option to return dict instead of FlatMapping from init, apply et al.
This change implements the functionality required to change the mapping type
returned from Haiku APIs from `FlatMapping` to `dict`.

FlatMapping was historically used for performance and immutability, but it is
not compatible with new "prefix based" APIs in JAX (e.g. `sharded_jit`). Recent
benchmarks show that the performance of dict is likely good enough (and perhaps
better for very small structures). Losing immutability in favour of using a type
from the standard library was felt to be a worthwhile tradeoff.

You can opt in early to this feature by setting the environment variable
`HAIKU_FLATMAPPING=0`.

PiperOrigin-RevId: 362479693
Change-Id: I7305dca73d9b65ccc484389b7d6ef15a171c7b24