Rename FlatMapping to FlatMap, override pickle.load for FlatMapping.
Old checkpoints refer to `haiku._src.data_structures.FlatMapping` as the
function to call to restore `FlatMapping` types (thanks to the previous
definition of `FlatMapping.__reduce__`). When `HAIKU_FLATMAPPING=0` we
would like to restore these values as `dict` (the Mapping type used by the rest
of Haiku).

The `FlatMapping` type now in Haiku is a shim, that on `__new__` defers to
`to_haiku_dict(..)` returning either a `FlatMap` or `dict`.

If users explicitly want FlatMap they can then call `to_immutable_dict` on the
restored value.

NOTE: We keep the binary format the FlatMap equivalent to FlatMapping at head.
PiperOrigin-RevId: 381002860