Provide a better error message for transform(jit(f)).
A very common issue (especially for new users) is to write something like the
following:

```python
@jax.jit
def f(x):
  # .. some stuff with haiku modules ..

f = hk.transform(f)
```

There is a subtle bug in this code, `f` (pre-transform) is not a pure function
(until after you hk.transform it). This means that JAX will (correctly) raise an
tracer error if you run the init or apply functions.

PiperOrigin-RevId: 409154814