Add `split_rng` argument to `hk.vmap(..)`.
`hk.vmap(..)` currently does not map any of the implicit state (params, state,
rng). From a users perspective, this means that rngs are effectively broadcast
within a vmap.

This has caused confusion for some users, who would expect the following to
produce 4 distinct keys (while it currently produces 4 identical keys):

    x = jnp.arange(4)
    hk.vmap(lambda _: hk.next_rng_key())(x)

To fix this, we propose an additional argument to vmap, allowing control over
how RNGs are split inside a vmap. This is inspired by the design in Flax, which
allows control over how RNGs are handled (as well as other hidden state).

>>> x = jnp.arange(4)
>>> hk.vmap(lambda _: hk.next_rng_key(), split_rng=True)(x)
DeviceArray([[1031272405,  285275113],
             [ 729498942, 3304897081],
             [3258709422, 3714857178],
             [  92606335, 1878054669]], dtype=uint32)

>>> hk.vmap(lambda _: hk.next_rng_key(), split_rng=False)(x)
DeviceArray([[255383827, 267815257],
             [255383827, 267815257],
             [255383827, 267815257],
             [255383827, 267815257]], dtype=uint32)

An important case that is poorly supported in this implementation is vmapping
the creation of module parameters. Currently `hk.vmap` is not sufficiently
expressive to allow you to say whether or not parameters are mapped. We hope to
address this in the future.

PiperOrigin-RevId: 420283486