Add a global and local way to disable generating automatic repr.
When Modules are constructed in Haiku, we generate a (hopefully!) useful string
to print in __{str,repr}__. The goal is to make it simple to see the constructor
arguments passed to the module, as well as the type etc.

This string serialisation can be slow if you have very large objects (e.g. you
are passing very large configs around) or the cost can add up significantly in
unit tests if you are creating and destroying a significant number of modules
and each one has a modest overhead.

Thus, we offer two APIs to disable this feature. Firstly, we have a global
option to disable this feature for all module instances:

>>> hk.experimental.module_auto_repr(False)
>>> str(hk.Linear(1))
'<...Linear object at ...>'

>>> hk.experimental.module_auto_repr(True)
>>> str(hk.Linear(1))
'Linear(output_size=1)'

Secondly, if you would like to retain this feature globally, but disable it for
specific subclasses you can do so by setting the AUTO_REPR attribute on the
subclass:

>>> class TakesAVeryLargeConfigModule(hk.Module):
...   AUTO_REPR = False
>>> str(TakesAVeryLargeConfigModule(my_config))
'<...TakesAVeryLargeConfigModule object at ...>'

PiperOrigin-RevId: 427966851