Require split_rng in `hk.vmap`.
This is a backwards incompatible change, but based on user feedback the legacy
default for this is surprising, so we are now asking users to explicitly pass a
value.

We know there are some cases where users import code they don't have access to
edit and that code may be using hk.vmap without passing split_rng. For those
users we support an opt-out mechanism where you can restore the default value:

>>> hk.vmap.require_split_rng = False
>>> k1, k2 = hk.vmap(hk.next_rng_key, in_size=2)()  # split_rng=False implied.
>>> assert (k1 == k2).all()

PiperOrigin-RevId: 428519466