Prepare for the new `jax.named_call` implementation.
Make jaxpr_info use the old named_call implementation.
We'll aim to support the new namestack implementation once it's the default.
jaxpr_info relies on the fact that named_call is a call primitive, which is
changing soon.

Under the new implementation, we don't actually raise the abstraction level,
so there will be no escaped tracers (hence the removed tests in base_test).

PiperOrigin-RevId: 445428447