Special case TensorFlow's _DictWrapper in check_mapping.
Since google/jax@cec717d3 jax2tf will now preserve `_DictWrapper` when used with
`jax.tree_utils.*`. This has a side effect of exposing more JAX user code to
`_DictWrapper`.

`_DictWrapper` is not a subclass of `MutableMapping` (although imho it should
be) and attempts to fix that led to metaclass conflict errors that it may take
some time to unpick. So for now we special case TF's `_DictWrapper` in
`check_mapping` to avoid providing an unhelpful error message.

PiperOrigin-RevId: 447171154