Fix confusing error when a parameter is missing + `without_apply_rng`.
The issue being that we init first, then check if the parameter is missing,
if init uses RNG, and we're not using apply_rng, a "RNG is None" error is
thrown instead of a "missing parameter" one.
The deeper issue is that we need to order things this way to support the
DO_NOT_STORE feature, which is allowed to re-initialise a parameter under
apply, and we can only know if that's the case when `init` has returned.
We can catch the MissingRNGError, and choose to raise a "missing param" error
instead.

PiperOrigin-RevId: 503180739