Plumb orig_class into MethodContext.
This allows easy access to the class that defined the method currently being
called. Note that when using inheritance this is not necessarily the same as
`type(context.module)` since you may be calling a method not overridden by the
subclass (in this case `orig_class` would be the class that defined the method).

PiperOrigin-RevId: 505108741