Add hk.experimental.{get_params,get_initial_state,get_current_state}.
These APIs allow introspection on copies of the intputs/outputs of init/apply,
allowing advanced users to provide better error messages (e.g. if a module
parameter is renamed, advanced users can write code that looks in the params dict
for parameters with the old name, and/or check the shape matches what is
expected).

PiperOrigin-RevId: 517124009