Add memory tiling to MultiHeadAttention
When the attn_logits tensor is expected to be very large (by
default above 1GB), attention calculation is split into chunks
sized below this limit.

This reduces memory consumption quite a bit and crucially allows
for input tensor shapes that would otherwise not fit in memory.
There also seems to be a small speed improvement, presumably due
to better caching.

This is a similar concept to FlashAttention, though a much
simpler implementation because it is only tiling in one dimension

As an example, calculating self attention with:

key_size = 64
num_heads = 16
h = jax.random.normal(jax.random.PRNGKey(42), [1,16000,1024])

This change resulted in:
a ~30% speedup (6s to 4.3s) and
a 5x reduction in memory usage (19.3GB to 3.9Gb)