fix version issues in docs build.
typing-extensions is no longer in our requirements.txt, but is still used.
We include typing-extensions transitively through our
requirements-txt dependencies, so all GH actions are succeeding (because they
all install the test requirements). But, our docs build does not,
and is failing as a consequence.

We no longer support python 3.7, so can import from typing instead.

PiperOrigin-RevId: 549567348