Fix `inspect.stack()` error caused by no_flax shim module.
Typically, `__file__` is set on a module when it is imported to the module's absolute path. However, when flax is not available Haiku injects a flax shim to still allow haiku to be imported. This causes issues when trying to e.g. `inspect.stack()`, as it will look for a `__file__` attr, which then raises an `ImportError`. Instead, we special case this attribute to return `None` to indicate it is a shim and not a real module.

PiperOrigin-RevId: 844747516